requirements.txt
aiogram==3.1.1
aiosqlite
httpx
python-dotenv